resolved infinite recursion in tests and seperated tests into repositorytests and minitwittests
