+++ java-itu-minitwit/src/test/java/RepositoriesTests.java
@@ -1,138 +1,9 @@
 import Logic.Logger;
-import Persistence.DB;
 import Persistence.Repositories;
-import RoP.Result;
-import RoP.Success;
 import Utilities.Hashing;
 import org.junit.jupiter.api.Test;
 
-import static spark.Spark.stop;
-
-class RepositoriesTest {
-    @org.junit.jupiter.api.BeforeEach
-    void setUp() {
-        DB.setDATABASE("testMinitwit");
-        if (System.getProperty("DB_TEST_CONNECTION_STRING") != null) {
-            DB.setCONNECTIONSTRING(System.getProperty("DB_TEST_CONNECTION_STRING"));
-            DB.setUSER(System.getProperty("DB_USER"));
-            DB.setPW(System.getProperty("DB_PASSWORD"));
-        }
-        Repositories.dropDB();
-    }
-
-    @org.junit.jupiter.api.AfterEach
-    void tearDown() {
-        stop();
-    }
-
-    //Helper functions
-
-    //Helper function to register a user
-    Result<String> register(String username, String password, String password2, String email){
-        if (password2==null) password2 = password;
-        if (email==null)     email = username + "@example.com";
-        return register(username, email, password, password2);
-    }
-
-    //Helper function to login
-    Result<Boolean> login(String username, String password) {
-        return Repositories.queryLogin(username, password);
-    }
-
-    //Helper function to register and login in one go
-    Result<Boolean> register_and_login(String username, String password) {
-        register(username, password, null, null);
-        return login(username, password);
-    }
-    int userId = 1;
-    //Helper function to register, login and get the id
-    Result<Integer> register_login_getID(String username, String password, String password2, String email) {
-        register(username, password, password2, email);
-        login(username, password);
-        var id = Repositories.getUserId(username);
-        assert (id.isSuccess());
-        assert (id.get() == userId);
-        userId++;
-        return id;
-    }
-    //Helper function to logout
-    Result<Boolean> logout() {
-        //todo logout helper function
-        return new Success<>(true);
-    }
-
-    //Records a message
-    void add_message(String text, int loggedInUserId) {
-        var rs = Repositories.addMessage(text, loggedInUserId);
-        assert (rs.get());
-        try {
-            Thread.sleep(100);      //hotfix: added to ensure order of messages
-        } catch (InterruptedException e) { }
-    }
-
-    //Tests
-
-
-    @Test
-    void test_register(){
-        Logger.processUsers();
-        assert ((int) Logger.getUsers() == 0);
-        assert (Repositories.countUsers().get() == 0);
-
-        var error = register("user1", "q123", null, null);
-        assert (error.isSuccess() && error.get().equals("OK"));
-        assert (Repositories.countUsers().get() == 1);
-        Logger.processUsers();
-        assert ((int) Logger.getUsers() == 1);
-
-        error = register("user1", "q123", null, null);
-        assert (!error.isSuccess() && error.getFailureMessage().equals("The username is already taken"));
-        error = register("", "q123", null, null);
-        assert (!error.isSuccess() && error.getFailureMessage().equals("You have to enter a username"));
-        error = register("user2", "", null, null);
-        assert (!error.isSuccess() && error.getFailureMessage().equals("You have to enter a password"));
-        error = register("user2", "1", "2", null);
-        assert (!error.isSuccess() && error.getFailureMessage().equals("The two passwords do not match"));
-        error = register("user2", "1", null, "bad email");
-        assert (!error.isSuccess() && error.getFailureMessage().equals("You have to enter a valid email address"));
-        assert (Repositories.countUsers().get() == 1);
-        Logger.processUsers();
-        assert ((int) Logger.getUsers() == 1);
-    }
-
-    @Test
-    void test_login_logout() {
-        var result = register_and_login("user1", "default");
-        assert (result.isSuccess() && result.get());
-        result = logout();
-        assert (result.isSuccess() && result.get()); //TODO will always succeed as is now
-        result = login("user2", "wrongpassword");
-        var msg = result.getFailureMessage();
-        assert (!result.isSuccess() && result.getFailureMessage().equals("Invalid username"));
-        result = login("user1", "wrongpassword");
-        assert (!result.isSuccess() && result.getFailureMessage().equals("Invalid password"));
-    }
-
-    @Test
-    void test_publicTimeline() {
-        var id1 = register_login_getID("foo", "default", null, null);
-
-        String text1 = "test message 1", text2 = "<test message 2>";
-        add_message(text1, id1.get());
-        add_message(text2, id1.get());
-        var rs = Repositories.publicTimeline();
-        assert (rs.isSuccess());
-        var tweet1 = rs.get().get(1);
-        var tweet2 = rs.get().get(0);
-        assert (tweet1.getEmail().equals("foo@example.com"));
-        assert (tweet1.getUsername().equals("foo"));
-        assert (tweet1.getText().equals(text1));
-
-        assert (tweet2.getEmail().equals("foo@example.com"));
-        assert (tweet2.getUsername().equals("foo"));
-        assert (tweet2.getText().equals(text2));//todo store as: "&lt;test message 2&gt;"
-    }
-
+class RepositoriesTests extends DatabaseTestBase{
     @Test
     void test_getPersonalTweetsById() {
         assert (Repositories.countMessages().get() == 0);
