+++ java-itu-minitwit/src/test/java/UtilitiesTests.java
@@ -1,0 +1,2 @@
+public class UtilitiesTests {
+}
